Give up expecting echo_client to be unicode safe
The example echo_client.py outputs in UTF-8 in Python 2 and the locale
encoding in Python 3. test_example_echo_client expects output to be
UTF-8. It's common (usual?) on Windows for the DOS codepage to be
non-UTF-8, so this causes the test to fail.

In general, echo_client.py won't be able to output arbitrary unicode
everywhere, so testing for it is a bit futile anyway.

Only test ASCII characters.