Give up expecting echo_client to be unicode safe (#23)
The example echo_client.py output in UTF-8 in Python 2 and the locale
encoding in Python 3. test_example_echo_client expected output to be
UTF-8. It's common (usual?) on Windows for the DOS codepage to be
non-UTF-8, so this caused the test to fail.

Change it to use the locale encoding everywhere. This means that we
aren't guaranteed to be able to output arbitrary unicode, so stop
testing for it.

Only test ASCII characters.